requirements.txt
adbutils==2.6.0
loguru==0.7.2
PyYAML==6.0.1
requests==2.31.0

